core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "redac_core"
version = "0.1.0"
description = "Core logic for RedactX PII redaction tool"
authors = [{name = "Your Name", email = "your.email@example.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openmed>=0.5.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "pymupdf>=1.20.0",
    "python-docx>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["redac_core"]
